fix: split stats rpc handlers
